+++ LMS.Entities/Question.cs
@@ -3,7 +3,7 @@
     public class Question
     {
         public int Id { get; set; }
-        public double Complexity { get; set; }
+        public int Complexity { get; set; }
         public string Content { get; set; }
 
         public bool IsVisible { get; set; }
